% Generated by roxygen2 (4.0.1): do not edit by hand
\name{r2vcf}
\alias{r2vcf}
\title{converts a list structure created by vcf2r back onto a VCF text file}
\usage{
r2vcf(y, fname)
}
\arguments{
\item{y}{is a list with a $head and a $body}

\item{fname}{is the file name}
}
\description{
converts a list structure created by vcf2r back onto a VCF text file
}
\seealso{
\code{\link{vcf2r}} which does the opposite
}

