% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vcf2r}
\alias{vcf2r}
\title{converts a VCF text file into an R structure with a head and a body}
\usage{
vcf2r(fname)
}
\arguments{
\item{fname}{is the file name}
}
\description{
converts a VCF text file into an R structure with a head and a body
}
\seealso{
\code{\link{r2vcf}} does the opposite
}

